% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Sum the square of difference between mean and your sample}
\usage{
sum_squares(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
Sum the square of difference between mean and your sample
}
\examples{
sum_squares(rnorm(100,20,2))
}
